Progress Made.
FSM is now active and working. stuff needs to be added, but the basis is there.

Deleting of nodes and the adding of nodes still needs to be added, as well as Wander needs fixing. My enemy's need to A* only if they cannot see the player. (possible Blackboard idea?) need to incorporate blackboard and utility AI, as well as other stuff.