wander for dumb enemy done
wander for dumb ai done. need seek and flee to be fixed for the dumb ai. as well as getting the utilityAI and blackboard for him done. then i need to implement collision avoidance and flocking for the smartAI, as well as fixing the a* to take 3 behaviours rather then 2.

6 days left.